refactor: move Content from adapter to content
